Move TestOut and TestResolver to test helpers
