+++ tests/export.test.ts
@@ -2,7 +2,7 @@ import test from 'ava'
 import { testEnvironment } from './helpers/test_environment'
 import exportCommand from '../src/commands/export'
 import { systemAPIEndpoint, graphcoolProjectFileName, graphcoolConfigFilePath } from '../src/utils/constants'
-import TestOut from '../src/system/TestOut'
+import TestOut from './helpers/test_out'
 import { mockedExportResponse } from './fixtures/mock_data'
 import { mockedPullProjectFile1 } from './fixtures/mock_data'
 import 'isomorphic-fetch'
