+++ tests/helpers/test_out.ts
@@ -1,7 +1,7 @@
-import { Out } from '../types'
+import { Out } from '../../src/types'
 import * as fs from 'fs'
 import * as chalk from 'chalk'
-import {setDebugMessage, contactUsInSlackMessage} from '../utils/constants'
+import {setDebugMessage, contactUsInSlackMessage} from '../../src/utils/constants'
 import figures = require('figures')
 
 export default class TestOut implements Out {
