+++ tests/push.test.ts
@@ -8,7 +8,7 @@ import {
   mockedPushSchemaResponse, mockProjectFile2, mockProjectFile3,
   mockedPullProjectResponse1, mockedPushSchemaResponseError, mockModifiedProjectFile1, mockedPushSchema1ResponseError
 } from './fixtures/mock_data'
-import TestOut from '../src/system/TestOut'
+import TestOut from './helpers/test_out'
 
 const fetchMock = require('fetch-mock')
 const debug = require('debug')('graphcool')
