+++ tests/utils_config.test.ts
@@ -1,6 +1,6 @@
 import test from 'ava'
-import TestResolver from '../src/system/TestResolver'
-import TestOut from '../src/system/TestOut'
+import TestResolver from './helpers/test_resolver'
+import TestOut from './helpers/test_out'
 import { Config } from '../src/utils/config'
 import { graphcoolConfigFilePath } from '../src/utils/constants'
 const configFileContent = '{"token":"1234"}'
